### 2018-01-08 #526 @whedon
[PRE REVIEW]: PyDMD: Python Dynamic Mode Decomposition
**Submitting author:** @mtezzele (<a href="http://orcid.org/0000-0001-9747-6328">Marco Tezzele</a>)
**Repository:** <a href="https://github.com/mathLab/PyDMD" target ="_blank">https://github.com/mathLab/PyDMD</a>
**Version:** v0.1.0
