### 2018-01-08 #526 @arfon
👋 @jakevdp - the submitting author suggested you as the handling editor for this submission.
