+++ joss-paper/paper.md
@@ -1,25 +1,25 @@
 ---
 title: 'PyDMD: Python Dynamic Mode Decomposition'
 tags:
-- Dynamic mode decomposition
-- DMD
-- Multiresolution DMD
-- Compressed DMD
-- Forward Backward DMD
+ - Dynamic mode decomposition
+ - DMD
+ - Multiresolution DMD
+ - Compressed DMD
+ - Forward Backward DMD
 authors:
-- name: Nicola Demo
- orcid: 0000-0003-3107-9738
- affiliation: 1
-- name: Marco Tezzele
- orcid: 0000-0001-9747-6328
- affiliation: 1
-- name: Gianluigi Rozza
- orcid: 0000-0002-0810-8812
- affiliation: 1
+ - name: Nicola Demo
+   orcid: 0000-0003-3107-9738
+   affiliation: 1
+ - name: Marco Tezzele
+   orcid: 0000-0001-9747-6328
+   affiliation: 1
+ - name: Gianluigi Rozza
+   orcid: 0000-0002-0810-8812
+   affiliation: 1
 affiliations:
-- name: Internation School of Advanced Studies, SISSA, Trieste, Italy
- index: 1
-date: - January 2018
+ - name: Internation School of Advanced Studies, SISSA, Trieste, Italy
+   index: 1
+date: 8 January 2018
 bibliography: paper.bib
 ---
 
