### 2018-01-08 #526 @arfon
@mtezzele - I made some small edits to your `paper.md` to (hopefully) fix things up: https://github.com/mathLab/PyDMD/pull/55
