+++ joss-paper/paper.bib
@@ -7,7 +7,16 @@
 
 %% Saved with string encoding Unicode (UTF-8) 
 
-
+@article{schmid2010dynamic,
+	Author = {Schmid, Peter J},
+	Date-Added = {2018-01-08 16:15:45 +0000},
+	Date-Modified = {2018-01-08 16:15:45 +0000},
+	Journal = {Journal of fluid mechanics},
+	Pages = {5--28},
+	Publisher = {Cambridge University Press},
+	Title = {Dynamic mode decomposition of numerical and experimental data},
+	Volume = {656},
+	Year = {2010}}
 
 @article{koopman1931hamiltonian,
 	Author = {Koopman, Bernard O},
@@ -32,5 +41,5 @@
 @url{pydmd,
 	Date-Added = {2018-01-04 14:16:03 +0000},
 	Date-Modified = {2018-01-04 14:16:03 +0000},
-	Title = {{PyDMD: Python Dynamic Mode Decomposition. Available at}: \url{https://github.com/mathLab/PyDMD}},
+	Title = {{PyDMD: Python Dynamic Mode Decomposition. Available at: \url{https://github.com/mathLab/PyDMD}}},
 	Urldate = {2018-01}}
