### 2018-01-10 #526 @arfon
> Reading the list I would suggest nicoguaro or lheagy. Is that something I need to choose or is it up to the editor? Sorry but I am new to this process...

This is something that the editor will help with.
