### 2018-01-11 #526 @jakevdp
I've reached out to a potential reviewer.
