### 2018-01-25 #530 @jkmacc-LANL
PyDMD is a well-documented, well-tested, easy to install software package.  Congratulations!  It was a pleasure to review this (...and thanks for your patience).  I only had two minor requests (below).  These aren't show-stoppers, so I didn't create issues for them in the project repo.  With these minor tweaks, this contribution is easily ready to be accepted.

## Functionality

I installed and ran the tutorial notebooks on OS X 10.12.6 with Python 3.6.4.  It went very smoothly.

* [ ] Tutorial notebook 2 requires `ffmpeg`.  I recommended that this be explicit at the top of the notebook, or in the cell (7) that uses it.

## Documentation

* [ ] A slightly longer description of the software on the documentation website landing page would be helpful.  Some of the text from the JOSS paper would be appropriate.

## Software Paper

* [ ] Please add DOIs to the references, where possible.
