### 2018-01-25 #530 @mtezzele
@jkmacc-LANL thank you very much! I will proceed in implementing your suggestions in the next few days!
