### 2018-01-29 #530 @mtezzele
@jakevdp I have implemented all the suggestions and regenerated the pdf. What do I have to do now?
