### 2018-01-29 #530 @jkmacc-LANL
@mtezzele I see your changes, and they look good.  Regarding the middle point, my suggestion was to add a slightly longer description to https://mathlab.github.io/PyDMD/, not on the README.  I apologize for the confusion.  Everything else is great👍
