### 2018-01-30 #530 @jakevdp
@jkmacc-LANL, has @mtezzele addressed all your comments?
