### 2018-02-05 #530 @jkmacc-LANL
Yes, they are addressed.  Thanks!
