### 2018-02-12 #530 @ndem0
@arfon We have upload the package and this is the DOI for all the versions:
 [https://doi.org/10.5281/zenodo.1171787](https://doi.org/10.5281/zenodo.1171787)
