### 2018-02-12 #530 @arfon
@ndem0 @mtezzele - your paper is now accepted in JOSS and your DOI is https://doi.org/10.21105/joss.00530 ⚡️ 🚀 💥
